Refactor Settings - extract AccountSettings
